backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0
aiosqlite==0.19.0
openai==1.12.0
httpx==0.25.2
sse-starlette==1.6.5
sqlmodel==0.0.16